refs #284 - added API runtime for RPC structure fields - Trail support for structures - Use length as Integer and decimalPlace field instead double - Handle sending of complex field (Not done yet) -

+++ openlegacy-designtime/src/main/java/org/openlegacy/designtime/rpc/source/parsers/RpcFieldDefinitionBuilder.java
@@ -20,7 +20,7 @@ public class RpcFieldDefinitionBuilder {
 		SimpleRpcFieldDefinition rpcFieldDefinition = new SimpleRpcFieldDefinition(javaFieldName, General.class);
 		rpcFieldDefinition.setOriginalName(fieldName);
 
-		rpcFieldDefinition.setLength(fieldFormater.getLength());
+		rpcFieldDefinition.setLength((int)fieldFormater.getLength());
 		rpcFieldDefinition.setJavaType(fieldFormater.getJavaType());
 		rpcFieldDefinition.setFieldTypeDefinition(fieldFormater.getType());
 		return rpcFieldDefinition;
